DataVisualizationAgent/src/DataVisualizationAgent/agent.py

import matplotlib.pyplot as plt
from autogen import AssistantAgent
from agoric_sdk import ContractAPI

class DataVisualizationAgent:
    def __init__(self):
        self.assistant = AssistantAgent("visualization_assistant")
        self.contract_api = ContractAPI("https://api.agoric.net")

    def collect_data(self, contract_address):
        try:
            # Collect data from the specified contract
            data = self.contract_api.get_contract_data(contract_address)
            return data
        except Exception as e:
            return f"Failed to collect data: {str(e)}"

    def analyze_data(self, data):
        try:
            # Perform data analysis
            analysis_result = self.assistant.handle_input(f"Analyze the following data: {data}")
            return analysis_result
        except Exception as e:
            return f"Failed to analyze data: {str(e)}"

    def visualize_data(self, data):
        try:
            # Visualize the data
            plt.figure(figsize=(10, 6))
            plt.plot(data['x'], data['y'], marker='o')
            plt.title('Data Visualization')
            plt.xlabel('X-axis')
            plt.ylabel('Y-axis')
            plt.grid(True)
            plt.savefig('data_visualization.png')
            return "Data visualization saved as data_visualization.png"
        except Exception as e:
            return f"Failed to visualize data: {str(e)}"

if __name__ == "__main__":
    agent = DataVisualizationAgent()
    contract_address = "example_contract_address"
    data = agent.collect_data(contract_address)
    print(data)
    analysis_result = agent.analyze_data(data)
    print(analysis_result)
    visualization_result = agent.visualize_data(data)
    print(visualization_result)
